Add some helper functions
The functions before work really well for quick asserts here and there.
When these functions are used as the only means of testing,
it becomes harder to know what caused the failure.

If we provide just a small bit of context (actual/expected),
we can make it quite a bit easier to understand failures.